requirements.txt
configobj==4.7.2
enum34==1.1.6
gevent==0.13.6
greenlet==0.3.1
httplib2==0.9.2
lxml==2.3.2
paho-mqtt==1.3.0
requests==2.7.0
